% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Example-class}
\alias{Example-class}
\title{Create Example object}
\description{
Create Example object
}
\section{Slots}{

\describe{
\item{\code{type}}{character}

\item{\code{expr_filename}}{character}

\item{\code{label_filename}}{character}
}}

